completed loan apply, get loan details for user ,get pending loan for employees,approve loan,reject loan,emiCalculator,EmiScheduler,Security Configurations

+++ backend/banking-service/src/main/java/com/finix/loan/repository/LoanTypeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.finix.loan.entity.LoanType;
 
 @Repository
-public interface LoanTypeRepository extends JpaRepository<LoanType, Integer>{
+public interface LoanTypeRepository extends JpaRepository<LoanType, Long>{
 
 }
